aga_knowledge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aga-knowledge"
version = "0.3.0"
description = "AGA Knowledge — 知识管理系统：知识注册、编码器对齐、混合检索（HNSW+BM25+RRF）、文档分片、图片处理、持久化、Portal API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "AGA Team"},
]
keywords = ["llm", "knowledge-management", "portal", "api", "aga"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = []

[project.optional-dependencies]
portal = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
]
redis = ["redis>=5.0.0"]
postgres = ["asyncpg>=0.28.0"]
yaml = ["pyyaml>=6.0"]
encoder = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
retrieval = [
    "hnswlib>=0.7.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0",
]
all = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "asyncpg>=0.28.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "hnswlib>=0.7.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/aga-project/aga-knowledge"
Documentation = "https://aga-project.github.io/docs"
Repository = "https://github.com/aga-project/aga-knowledge"

[project.scripts]
aga-portal = "aga_knowledge.portal.app:main"

[tool.setuptools.packages.find]
include = ["aga_knowledge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
